Move atoms to themed components
