+++ frontend/src/features/theme/components/SectionTitle.tsx
